requirement.txt
Pillow
cmake == 3.26.4
dlib
face-recognition == 1.3.0
face-recognition-models == 0.3.0
opencv-contrib-python == 4.9.0.80
pyfiglet == 1.0.2
gTTS
pygame
